Allow custom item colours
This commit adds the ability to change the background and text colours of individual calendar entries. Additionally, the colours can be reverted back to the default. These options are accessed by right clicking an item or navigating to the "Item" menu when an item is selected.

This commit addresses issue #5.